requirements.txt
factory_boy
pytest
pytest-cov
